+++ src/main/java/no/elg/ii/inventory/slot/InventorySlot.java
@@ -27,9 +27,6 @@
 
 package no.elg.ii.inventory.slot;
 
-import javax.annotation.Nonnull;
-import net.runelite.api.widgets.Widget;
-
 public interface InventorySlot {
 
   /**
@@ -58,13 +55,6 @@ public interface InventorySlot {
 
   int getQuantity();
 
-  /**
-   * @param widget
-   * @return If the item id or quantity is different from this slot
-   */
-  default boolean isDifferent(@Nonnull Widget widget) {
-    return getItemId() != widget.getItemId() || getQuantity() != widget.getItemQuantity();
-  }
 
   /**
    * @return Whether this slot is valid, i.e. has an item id