+++ src/main/java/no/elg/ii/InstantInventoryConfig.java
@@ -54,6 +54,23 @@ public interface InstantInventoryConfig extends Config {
   )
   String FEATURE_SECTION = "instant-inventory-features";
 
+  @ConfigSection(
+    name = "Common",
+    description = "Common settings for all features",
+    position = 10
+  )
+  String COMMON_SECTION = "instant-inventory-common";
+
+  @ConfigSection(
+    name = "Advanced",
+    description = "Advanced settings not recommended to change unless you know what you are doing",
+    position = 100,
+    closedByDefault = true
+  )
+  String ADVANCED_SECTION = "instant-inventory-advanced";
+
+  ///////////////////////////////////////////////////////////////////////////////////////////////////////////////
+
   @ConfigItem(
     keyName = DROP_CONFIG_KEY,
     name = "Drop Items Instantly",
@@ -118,6 +135,7 @@ public interface InstantInventoryConfig extends Config {
 
 
   @ConfigItem(
+    section = COMMON_SECTION,
     keyName = "maxUnmodifiedTicks",
     name = "Max Unmodified Ticks",
     description =
@@ -136,6 +154,7 @@ public interface InstantInventoryConfig extends Config {
   }
 
   @ConfigItem(
+    section = COMMON_SECTION,
     keyName = "hideOpacity",
     name = "Hide opacity",
     description = "A number between " + WidgetUtil.FULLY_OPAQUE + " and " + WidgetUtil.FULLY_TRANSPARENT + "." +
@@ -150,6 +169,7 @@ public interface InstantInventoryConfig extends Config {
   }
 
   @ConfigItem(
+    section = COMMON_SECTION,
     keyName = "changeOpacity",
     name = "Change opacity",
     description = "A number between " + WidgetUtil.FULLY_OPAQUE + " (fully visible) and " + WidgetUtil.FULLY_TRANSPARENT + "(fully hidden)." +
@@ -162,11 +182,17 @@ public interface InstantInventoryConfig extends Config {
     return WidgetUtil.HALF_TRANSPARENT;
   }
 
+
+  ///////////////////////////////////////////////////////////////////////////////////////////////////////////////
+
+
   @ConfigItem(
+    section = ADVANCED_SECTION,
     keyName = "allowWidgetForcing",
     name = "Allow Widget Forcing",
-    description = "Force the state of widgets to the predicted next state." +
-      "<p>This allows for a more seamless experience, but may cause problems as the client may update the widgets after the plugin has done so.",
+    description =
+      "Force the state of widgets to the predicted next state." +
+        "<p>This allows for a more seamless experience, but may cause problems as the client may update the widgets after the plugin has done so.",
     warning = "Disabling this will cause the clicked widgets to not be properly updated.<p>Only disable if there is issues with the plugin.",
     position = 120
   )