+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+* Fix the quantity of the item in the bank not being updated client-side when depositing items
 * Fix missing unit in config
 
 ### Removed